requirements.txt
google-genai
requests
feedparser
python-dateutil
python-dotenv
